@@ -1,49 +1,47 @@
 ﻿using System.Threading.Tasks;
-using RDNET.Test;
 using Xunit;
 
-namespace AllDebridNET.Test
+namespace AllDebridNET.Test;
+
+public class HostsTest
 {
-    public class HostsTest
+    [Fact]
+    public async Task Hosts()
     {
-        [Fact]
-        public async Task Hosts()
-        {
-            var client = new AllDebridNETClient("AllDebridNETTest", Setup.ApiKey);
+        var client = new AllDebridNETClient("AllDebridNETTest", Setup.ApiKey);
 
-            var result = await client.Hosts.GetHostsAsync();
+        var result = await client.Hosts.GetHostsAsync();
 
-            Assert.True(result.HostList.Count > 50);
-        }
+        Assert.True(result.HostList.Count > 50);
+    }
 
-        [Fact]
-        public async Task Domains()
-        {
-            var client = new AllDebridNETClient("AllDebridNETTest", Setup.ApiKey);
+    [Fact]
+    public async Task Domains()
+    {
+        var client = new AllDebridNETClient("AllDebridNETTest", Setup.ApiKey);
 
-            var result = await client.Hosts.GetDomainsAsync();
+        var result = await client.Hosts.GetDomainsAsync();
 
-            Assert.True(result.Hosts.Count > 50);
-        }
+        Assert.True(result.Hosts.Count > 50);
+    }
         
-        [Fact]
-        public async Task HostPriority()
-        {
-            var client = new AllDebridNETClient("AllDebridNETTest", Setup.ApiKey);
+    [Fact]
+    public async Task HostPriority()
+    {
+        var client = new AllDebridNETClient("AllDebridNETTest", Setup.ApiKey);
 
-            var result = await client.Hosts.GetHostsPriorityAsync();
+        var result = await client.Hosts.GetHostsPriorityAsync();
 
-            Assert.True(result.Count > 50);
-        }
+        Assert.True(result.Count > 50);
+    }
         
-        [Fact]
-        public async Task HostsForUser()
-        {
-            var client = new AllDebridNETClient("AllDebridNETTest", Setup.ApiKey);
+    [Fact]
+    public async Task HostsForUser()
+    {
+        var client = new AllDebridNETClient("AllDebridNETTest", Setup.ApiKey);
 
-            var result = await client.Hosts.GetHostsForUserAsync();
+        var result = await client.Hosts.GetHostsForUserAsync();
 
-            Assert.True(result.HostList.Count > 50);
-        }
+        Assert.True(result.HostList.Count > 50);
     }
 }
